+++ src/main/java/com/practicespring/printerstore/exceptions/ProductNotFoundException.java
@@ -1,7 +1,7 @@
 package com.practicespring.printerstore.exceptions;
 
 public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(int id) {
+    public ProductNotFoundException(Long id) {
         super("Could not found product " + id);
     }
 }
